[NLPCC-UI 2025] Final Project - Syaqila Ghavaro Adista, Bernadethe Impresa wangge/requirement.txt
Flask==2.3.3
google-generativeai==0.3.2
google-api-python-client==2.108.0
requests==2.31.0
python-dotenv==1.0.0